Task DEV 6 added a tester class and a description of the Custom Stack methods.
